% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{error}
\alias{error}
\title{Log an Error Message}
\usage{
error(
  message,
  ...,
  body = NA_character_,
  target = "shiny",
  event_type = "log",
  parent = NA_character_,
  .envir = parent.frame()
)
}
\arguments{
\item{message}{A character string that can be processed by \code{glue::glue()}.}

\item{...}{Arguments to be passed to \code{glue::glue()} for interpolation into \code{message}.}

\item{body}{An optional body for the log event. Defaults to \code{NA_character_}.}

\item{target}{The target of the log event. Defaults to \code{"shiny"}.}

\item{event_type}{The type of the event. Defaults to \code{"log"}.}

\item{parent}{The ID of a parent event, if any. Defaults to \code{NA_character_}.}

\item{.envir}{The environment in which to evaluate the glue expressions in \code{message}.}
}
\value{
The generated UUID for the event, as a character string.
}
\description{
A wrapper around \code{\link[=log]{log()}} for error messages (level = "Error").
}
\seealso{
\code{\link[=log]{log()}}
}
