% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{get_session_id}
\alias{get_session_id}
\title{Get Session ID}
\usage{
get_session_id(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The shiny session object. Defaults to \code{shiny::getDefaultReactiveDomain()}.}
}
\value{
The session ID as a character string.
}
\description{
Retrieves the Faucet session ID from the Shiny session's request object.
}
